refactor: extract SerialFraming from SerialAdapter
Move the protobuf framing state machine (FrameState enum, frameState/
frameLength/frameIndex/payloadBuffer fields, encodeMeshMessage,
decodeMeshMessage, injectByte) out of SerialAdapter into a new
standalone SerialFraming class.

- SerialFraming::injectByte is always public (no #ifdef UNIT_TEST),
  with frameBuffer()/frameLen() accessors valid after it returns true
- SerialAdapter delegates to _framing member; loop() simplified to
  call _framing.injectByte(b)
- test_serial_framing.cpp extended with SerialFramingTest suite
  exercising encode/decode round-trip and injectByte state machine
  directly, without Mesh plumbing
