+++ main/src/adapter/serial/SerialAdapter.cpp
@@ -57,10 +57,8 @@ void SerialAdapter::sendHealthReport() {
 }
 
 SerialAdapter::SerialAdapter(int pin)
-    : Adapter(pin), frameState(FrameState::AwaitingLen1), frameLength(0), frameIndex(0),
-      lastReportedHopCount(0) {
+    : Adapter(pin), lastReportedHopCount(0) {
   _adapterType = adapter_types::SERIAL_ADAPTER;
-  memset(payloadBuffer, 0, sizeof(payloadBuffer));
 
   Logger::logln("Serial_Adapter", "Serial_Adapter constructed with pin " + String(pin),
                 LogLevel::LOG_INFO);
@@ -91,51 +89,8 @@ void SerialAdapter::loop() {
 
   while (Serial.available() > 0) {
     uint8_t byteIn = static_cast<uint8_t>(Serial.read());
-
-    switch (frameState) {
-    case FrameState::AwaitingLen1:
-      frameLength = byteIn;
-      frameState = FrameState::AwaitingLen2;
-      break;
-
-    case FrameState::AwaitingLen2:
-      frameLength |= static_cast<uint16_t>(byteIn) << 8;
-
-      if (frameLength == 0 || frameLength > MAX_PAYLOAD) {
-        Logger::logln("SERIAL", "Frame parse error", LogLevel::LOG_WARN);
-        lattice::err::fail(lattice::core::ErrorTypeDigit::COMM, lattice::core::ModuleDigit::ADAPTER,
-                           2, "Serial_Adapter: Invalid frame length");
-        // Reset on invalid length
-        frameState = FrameState::AwaitingLen1;
-        frameLength = 0;
-        frameIndex = 0;
-      } else {
-        frameIndex = 0;
-        frameState = FrameState::AwaitingPayload;
-      }
-      break;
-
-    case FrameState::AwaitingPayload:
-      if (frameIndex >= MAX_PAYLOAD) {
-        Logger::logln("SERIAL", "Frame parse error", LogLevel::LOG_WARN);
-        lattice::err::fail(lattice::core::ErrorTypeDigit::COMM, lattice::core::ModuleDigit::ADAPTER,
-                           3, "Serial_Adapter: Frame buffer overflow");
-        frameState = FrameState::AwaitingLen1;
-        frameLength = 0;
-        frameIndex = 0;
-        break;
-      }
-
-      payloadBuffer[frameIndex++] = byteIn;
-
-      if (frameIndex >= frameLength) {
-        LOG_D("SERIAL", "Frame complete %u bytes", frameLength);
-        handleCompleteFrame(payloadBuffer, frameLength);
-        frameState = FrameState::AwaitingLen1;
-        frameLength = 0;
-        frameIndex = 0;
-      }
-      break;
+    if (_framing.injectByte(byteIn)) {
+      handleCompleteFrame(_framing.frameBuffer(), _framing.frameLen());
     }
   }
 }
@@ -178,8 +133,8 @@ void SerialAdapter::onMeshDataImpl(const lattice::mesh::mesh_message& message) {
   }
 
   // Forward message to server via serial (existing encoding logic)
-  uint8_t encoded[MAX_PAYLOAD];
-  size_t n = encodeMeshMessage(message, encoded, sizeof(encoded));
+  uint8_t encoded[256];
+  size_t n = lattice::adapter::serial::SerialFraming::encode(message, encoded, sizeof(encoded));
 
   if (n == 0) {
     Logger::logln("Serial_Adapter", "Failed to encode mesh message for serial output",
@@ -201,105 +156,6 @@ void SerialAdapter::onMeshDataImpl(const lattice::mesh::mesh_message& message) {
   Logger::logln("Serial_Adapter", "Mesh message sent to serial successfully", LogLevel::LOG_DEBUG);
 }
 
-size_t SerialAdapter::encodeMeshMessage(const lattice::mesh::mesh_message& msg, uint8_t* out,
-                                        size_t outCap) {
-  Logger::logln("Serial_Adapter",
-                "Encoding mesh message - Type: " + String((uint8_t)msg.message_type) +
-                    " DataType: " + String(static_cast<int32_t>(msg.data_type)) +
-                    " HopCount: " + String(msg.hop_count),
-                LogLevel::LOG_DEBUG);
-
-  mesh_MeshMessage pbMsg = mesh_MeshMessage_init_zero;
-  pbMsg.messageType = static_cast<uint32_t>(msg.message_type);
-  pbMsg.dataType = static_cast<int32_t>(msg.data_type);
-  pbMsg.hopCount = msg.hop_count;
-  pbMsg.epochNum = msg.epoch_num;
-  pbMsg.seqNum = static_cast<uint32_t>(msg.seq_num);
-  pbMsg.protoVersion = static_cast<uint32_t>(msg.proto_version);
-  memcpy(pbMsg.originMacAddress, msg.origin_mac_address, 6);
-  memcpy(pbMsg.targetMacAddress, msg.target_mac_address, 6);
-  memcpy(pbMsg.lastHopMacAddress, msg.last_hop_mac_address, 6);
-
-  // data field: always present (12 bytes)
-  pbMsg.has_data = true;
-  pbMsg.data.size = sizeof(msg.data);
-  memcpy(pbMsg.data.bytes, msg.data, sizeof(msg.data));
-
-  // public_key: only encode for enrollment-related message types when non-zero
-  if (msg.message_type == MESH_TYPE_ENROLLMENT || msg.message_type == MESH_TYPE_JOIN_ACK) {
-    bool nonZero = false;
-    for (int i = 0; i < 32; ++i) {
-      if (msg.enrollment_public_key[i]) {
-        nonZero = true;
-        break;
-      }
-    }
-    if (nonZero) {
-      pbMsg.has_public_key = true;
-      pbMsg.public_key.size = 32;
-      memcpy(pbMsg.public_key.bytes, msg.enrollment_public_key, 32);
-    }
-  }
-
-  pb_ostream_t stream = pb_ostream_from_buffer(out, outCap);
-  if (!pb_encode(&stream, mesh_MeshMessage_fields, &pbMsg)) {
-    Logger::logln("Serial_Adapter", "nanopb encode failed", LogLevel::LOG_ERROR);
-    return 0;
-  }
-
-  Logger::logln("Serial_Adapter",
-                "Successfully encoded mesh message to " + String(stream.bytes_written) + " bytes",
-                LogLevel::LOG_DEBUG);
-  return stream.bytes_written;
-}
-
-bool SerialAdapter::decodeMeshMessage(const uint8_t* data, size_t len,
-                                      lattice::mesh::mesh_message& outMsg) {
-  Logger::logln("Serial_Adapter", "Decoding protobuf message of " + String(len) + " bytes",
-                LogLevel::LOG_DEBUG);
-
-  memset(&outMsg, 0, sizeof(outMsg));
-
-  mesh_MeshMessage pbMsg = mesh_MeshMessage_init_zero;
-  pb_istream_t stream = pb_istream_from_buffer(data, len);
-  if (!pb_decode(&stream, mesh_MeshMessage_fields, &pbMsg)) {
-    Logger::logln("Serial_Adapter", "nanopb decode failed", LogLevel::LOG_ERROR);
-    return false;
-  }
-
-  outMsg.message_type = static_cast<lattice::mesh::MeshMessageType>(pbMsg.messageType);
-  outMsg.data_type = static_cast<lattice::adapter::adapter_types>(pbMsg.dataType);
-  outMsg.hop_count = static_cast<uint8_t>(pbMsg.hopCount);
-  outMsg.epoch_num = pbMsg.epochNum;
-  outMsg.seq_num = static_cast<uint16_t>(pbMsg.seqNum);
-  outMsg.proto_version = static_cast<uint8_t>(pbMsg.protoVersion);
-  memcpy(outMsg.target_mac_address, pbMsg.targetMacAddress, 6);
-
-  if (pbMsg.has_data) {
-    size_t dataToCopy = pbMsg.data.size < 12u ? pbMsg.data.size : 12u;
-    memcpy(outMsg.data, pbMsg.data.bytes, dataToCopy);
-  }
-
-  if (pbMsg.has_public_key) {
-    memcpy(outMsg.enrollment_public_key, pbMsg.public_key.bytes, 32);
-  }
-
-  // For server-to-device messages (JOIN_ACK, SERIAL_CMD_BROADCAST) the MAC
-  // fields on the wire are meaningful and must not be overwritten.
-  // For device-originated relays (ADAPTER_DATA, MASTER_BEACON) the server
-  // leaves routing fields blank, so we fill them in with our own MAC.
-  if (outMsg.message_type != MESH_TYPE_JOIN_ACK &&
-      outMsg.message_type != MESH_TYPE_SERIAL_CMD_BROADCAST) {
-    readOwnMac(outMsg.origin_mac_address);
-    readOwnMac(outMsg.last_hop_mac_address);
-  } else {
-    memcpy(outMsg.origin_mac_address, pbMsg.originMacAddress, 6);
-    memcpy(outMsg.last_hop_mac_address, pbMsg.lastHopMacAddress, 6);
-  }
-
-  Logger::logln("Serial_Adapter", "Successfully decoded protobuf message", LogLevel::LOG_DEBUG);
-  return true;
-}
 
 void SerialAdapter::relayEnrollmentToServer(const uint8_t mac[6], const uint8_t pubKey[32]) {
   lattice::mesh::mesh_message msg = {};
@@ -309,7 +165,7 @@ void SerialAdapter::relayEnrollmentToServer(const uint8_t mac[6], const uint8_t 
   memcpy(msg.enrollment_public_key, pubKey, 32);
 
   uint8_t encoded[128];
-  size_t n = encodeMeshMessage(msg, encoded, sizeof(encoded));
+  size_t n = lattice::adapter::serial::SerialFraming::encode(msg, encoded, sizeof(encoded));
   if (n == 0) {
     Logger::logln("Serial_Adapter", "Failed to encode enrollment relay message",
                   LogLevel::LOG_ERROR);
@@ -367,7 +223,7 @@ void SerialAdapter::handleCompleteFrame(const uint8_t* data, size_t len) {
 #endif
 
   lattice::mesh::mesh_message msg;
-  if (!decodeMeshMessage(data, len, msg)) {
+  if (!lattice::adapter::serial::SerialFraming::decode(data, len, msg)) {
     Logger::logln("Serial_Adapter", "Failed to decode protobuf frame", LogLevel::LOG_ERROR);
     lattice::err::fail(lattice::core::ErrorTypeDigit::COMM, lattice::core::ModuleDigit::ADAPTER, 5,
                        "Serial_Adapter: Failed to decode protobuf frame");
@@ -523,49 +379,6 @@ void SerialAdapter::handleCompleteFrame(const uint8_t* data, size_t len) {
 
   Logger::logln("Serial_Adapter", "Frame processing completed successfully", LogLevel::LOG_DEBUG);
 }
-
-#ifdef UNIT_TEST
-bool SerialAdapter::injectByte(uint8_t byteIn) {
-  switch (frameState) {
-  case FrameState::AwaitingLen1:
-    frameLength = byteIn;
-    frameState = FrameState::AwaitingLen2;
-    break;
-
-  case FrameState::AwaitingLen2:
-    frameLength |= static_cast<uint16_t>(byteIn) << 8;
-    if (frameLength == 0 || frameLength > MAX_PAYLOAD) {
-      // Invalid length — reset
-      frameState = FrameState::AwaitingLen1;
-      frameLength = 0;
-      frameIndex = 0;
-    } else {
-      frameIndex = 0;
-      frameState = FrameState::AwaitingPayload;
-    }
-    break;
-
-  case FrameState::AwaitingPayload:
-    if (frameIndex >= MAX_PAYLOAD) {
-      frameState = FrameState::AwaitingLen1;
-      frameLength = 0;
-      frameIndex = 0;
-      break;
-    }
-    payloadBuffer[frameIndex++] = byteIn;
-    if (frameIndex >= frameLength) {
-      // Frame complete
-      _lastCompletedOpcode = payloadBuffer[0];
-      frameState = FrameState::AwaitingLen1;
-      frameLength = 0;
-      frameIndex = 0;
-      return true;
-    }
-    break;
-  }
-  return false;
-}
-#endif
 
 } // namespace adapter
 } // namespace lattice
